+++ packages-ui/roosterjs-react/lib/common/index.ts
@@ -5,4 +5,5 @@ export {
 } from './type/LocalizedStrings';
 export { default as UIUtilities } from './type/UIUtilities';
 export { default as ReactEditorPlugin } from './type/ReactEditorPlugin';
+export { default as createUIUtilities } from './utils/createUIUtilities';
 export { default as getLocalizedString } from './utils/getLocalizedString';
